move create shortlink to public_shares.go
